requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.0
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
aioredis==2.0.1

# Elasticsearch
elasticsearch==8.11.0

# Kafka
aiokafka==0.8.0
kafka-python==2.0.2

# Monitoring
prometheus-client==0.19.0
jaeger-client==4.8.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Email
python-dotenv==1.0.0
email-validator==2.1.0
aiohttp==3.9.1

# Utilities
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-dateutil==2.8.2
pytz==2023.3
uuid==1.30

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Video processing
ffmpeg-python==0.2.0
Pillow==10.1.0

# Analytics
clickhouse-driver==0.2.6
pandas==2.1.4
numpy==1.24.3

# Object Storage
minio==7.2.1
boto3==1.34.0

# CORS
fastapi-cors==0.1.0